Refactor submodule structure and imports
