+++ dask_cloudprovider/azure/setup/start_worker.py
