+++ dask_cloudprovider/digitalocean/tests/test_droplet.py
@@ -4,7 +4,7 @@ import dask
 
 digitalocean = pytest.importorskip("digitalocean")
 
-from dask_cloudprovider.providers.digitalocean.droplet import DropletCluster
+from dask_cloudprovider.digitalocean.droplet import DropletCluster
 from dask.distributed import Client
 from distributed.core import Status
