+++ dask_cloudprovider/generic/cloud-init.yaml.j2
